services/mcp_service/requirements.txt
# MCP Service dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
httpx>=0.24.0

# Service discovery
nacos-sdk-python>=0.1.14

# Configuration
python-dotenv>=1.0.0

# Tools
beautifulsoup4>=4.12.0
langchain-core>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
